Game Built
This game is finally in a basic state. Only 5 basic pieces right now, but the way I did the piece data should make it very simple to add any piece in the future. You can place the pieces if they are in valid spots, and the helper allows you to see what you are going to do. Pieces do not clear yet, and there is no scoring yet.